ezjs-extension index

Library ezjs-extension.browser

The entry point of this library is the module: Browser.

Library ezjs-extension.browser-common

This library exposes the following toplevel modules:

Browser_action_browser_common

I18n_browser_common

Runtime_browser_common

Storage_browser_common

Tabs_browser_common

Windows_browser_common

Library ezjs-extension.browser-lwt

The entry point of this library is the module: Browser_lwt.

Library ezjs-extension.chrome

The entry point of this library is the module: Chrome.

Library ezjs-extension.chrome-common

This library exposes the following toplevel modules:

Browser_action_chrome_common

I18n_chrome_common

Runtime_chrome_common

Storage_chrome_common

Tabs_chrome_common

Windows_chrome_common

Library ezjs-extension.chrome-lwt

The entry point of this library is the module: Chrome_lwt.

Library ezjs-extension.common

This library exposes the following toplevel modules:

Browser_utils

Runtime_utils

Tabs_utils

Library ezjs-extension.declarative-content

The entry point of this library is the module: Declarative_content.